backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
supabase==2.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
postgrest==0.16.2
gotrue==2.2.0
storage3==0.7.1
supafunc==0.4.1
